phenocam_EN

Data structure containing budburst dates for all evergreen needleleaf (EN) sites included in the first PhenoCam data release. These data are described in detail in Richardson et al. (2017)

phenocam_EN

Format

A structured list with 6 variables:

site

PhenoCam site name

location

geographic location

doy

DOY (1 - 365 as a vector)

transition_dates

budburst dates, as per 25 percent of the Gcc amplitude

Ti

mean daily temperature, matrix 365 rows x # site years columns

Li

daylength, matrix 365 rows x # site years columns