Fit multiple models for comparison

Model comparison routine to faciliate model development and quick comparisons of the skill of various models.

pr_fit_comparison(
  random_seeds = c(1, 12, 40),
  models = c("LIN", "TT", "TTs", "PTT", "PTTs", "M1", "M1s", "AT", "SQ", "SQb", "SM1",
    "SM1b", "PA", "PAb", "PM1", "PM1b", "UN", "UM1", "SGSI", "AGSI"),
  data = phenor::phenocam_DB,
  method = "GenSA",
  control = list(max.call = 10, temperature = 10000),
  par_ranges = system.file("extdata", "parameter_ranges.csv", package = "phenor",
    mustWork = TRUE),
  ncores = 1
)

Arguments

random_seeds

a vector with random seeds for cross validation

models

list of models to compare

data

which standard or custom dataset to use

method

optimization method to use (default = GenSA) - GenSA : Generalized Simulated Annealing algorithm - genoud : GENetic Optimization Using Derivatives - BayesianTools: various bayesian based optimization tools

control

additional optimization control parameters (default = list(max.call = 5000, temperature = 10000))

par_ranges

location of the parameter ranges of the models

ncores

number of cores to use to calculate model comparisons, system specific and defaults to 1 threat (default = 1)

Examples

# estimate will return the best estimated parameter set given the
# validation data
if (FALSE) {
my_comparison <- pr_fit_comparison(random_seeds = c(38,1),
 models = c("TT","PTT"),
 dataset = "phenocam_DB",
 par_ranges = "parameter_ranges.csv")
}