Formatting of Berkeley Earth Gridded temperature data

Formatting of Berkeley Earth Gridded temperature data

pr_fm_be(
  path = tempdir(),
  year = 2011,
  offset = 264,
  extent = c(-126, -66, 23, 54),
  internal = TRUE
)

Arguments

path

a path to the gridded data, only average temperature data will be considered

year

year to process (requires year - 1 to be present)

offset

offset of the time series in DOY (default = 264, sept 21)

extent

geographic coordinates constraining the output, defined as bottom left, top right c(lon1, lat1, lon2, lat2) (default = c(-126, -66, 23, 54))

internal

return results as an R variable or save as an RDS file

Examples

if (FALSE) { # \dontrun{
# run with default settings
# extracts values of the referenced publication
# see github README
be_data <- pr_fm_be(
 path = "/your/path",
 year = 2010)
} # }