Formatting Daymet tiled data

Formats Daymet tiled data into a format which can be ingested by the models in phenor

pr_fm_daymet_tiles(
  path = tempdir(),
  year = 2014,
  tile = 11935,
  offset = 264,
  internal = TRUE
)

Arguments

path
a path to tiled data

year
year to process (requires year - 1 to be present)

tile
daymet tile number

offset
offset of the time series in DOY (default = 264, sept 21)

internal
return results as an R variable or save as an RDS file

Value

Data file adhering to the phenor modelling input formatting. Data serves as input for model spatial model runs. The only required data input is the mean daily temperature, precipitation and VPD can be included but are not mandatory. Depending on the models used, including only temperature will keep file sizes down and processing time lower.

Examples

# run with default settings
# looks for daymet average temperature
# data in your home directory. These data
# are calculated using daymet_tmean() from
# the daymetr package
if (FALSE) {
daymet_data <- pr_fm_daymet_tiles()
}