Formatting E-OBS data

Formats E-OBS data as described by: Haylock, M.R., N. Hofstra, A.M.G. Klein Tank, E.J. Klok, P.D. Jones, M. New. 2008: A European daily high-resolution gridded dataset of surface temperature and precipitation. J. Geophys. Res (Atmospheres), 113, D20119, into a 'phenor' compatible list.

pr_fm_eobs(
  path = tempdir(),
  year = 2014,
  offset = 264,
  resolution = 0.25,
  internal = TRUE
)

Arguments

path

a path of the gridded data netCDF files

year

year to process (requires year - 1 to be present / downloaded)

offset

offset of the time series in DOY (default = 264, sept 21)

resolution

0.25 or 0.50 degree data, only the normal rectangular data will be processed not the polar variety. calculation overhead

internal

TRUE / FALSE, write data structure to file or not (as .rds)

Value

Returns spatial model input data from long term E-OBS modelled data. This data can be run by models specified in the phenor package.

Examples

if (FALSE) {
# run with default settings netCDF E-OBS files will be
# looked for in your home directory
eobs_data <- pr_fm_eobs()
}