Formatting PEP725 data

Some pre-processing steps are required as downloading the PEP725. So, for the species of interest download the separate zipped files. Unzip all files and put the respective scientific data in one folder.

pr_fm_pep725(
  pep_path = tempdir(),
  eobs_path = tempdir(),
  bbch = "11",
  species = NULL,
  offset = 264,
  count = 60,
  resolution = 0.25,
  pep_data
)

Arguments

pep_path

path to the PEP725 data (species files only)

eobs_path

path to regular grid E-OBS data.

bbch

which phenophase (bbch) to use (default = 11)

species

species to select from merged PEP725 file

offset

offset of the time series in DOY (default = 264, sept 21)

count

minimum number of acquisitions per location

resolution

resolution of the E-OBS data (0.25 or 0.5, default = 0.25)

pep_data

prefiltered subset of the merged PEP725 data obtained through merge_PEP725(), else the complete data contained in pep_path will be used

Value

returns a nested list of site locations, their respective phenological metrics and matching environmental data as extracted from the E-OBS product (corrected for altitude using a lapse rate of 5C/km.)

Details

The routine requires E-OBS climate, which can be downloaded from the E-OBS website and should be put unzipped in a single folder. (http://www.ecad.eu/download/ensembles/ensembles.php).

BBCH list: https://www.reterurale.it/downloads/BBCH_engl_2001.pdf

Examples

# run with default settings
# looks for transition date files derived
# through phenocamr in your home directory
# change the path to match your setup
if (FALSE) {
phenocam_data = pr_fm_pep725(pep_path = "~/pep_data/",
                              eobs_path = "~/eobs_data/")
}